refactor: complete networks/ package split (5/N)
