+++ stisim/networks/msm.py
@@ -1,0 +1,114 @@
+"""Men-who-have-sex-with-men (MSM) sexual networks."""
+import numpy as np
+import starsim as ss
+from .base import NoPartnersFound
+from .mf import MFNetwork
+
+__all__ = ['AgeMatchedMSM', 'AgeApproxMSM']
+
+
+class AgeMatchedMSM(MFNetwork):
+    """Men-who-have-sex-with-men network using exact age-sorted matching.
+
+    Extends :class:`StructuredSexual` for MSM partnerships. Eligible males
+    are sorted by age and paired sequentially so that partners have similar
+    ages. The ``msm_share`` parameter controls what fraction of males
+    participate.
+
+    Args:
+        pars (dict): Parameter overrides; key parameter is ``msm_share``
+            (default ``ss.bernoulli(p=0.015)``).
+        **kwargs: Additional parameter overrides.
+    """
+
+    def __init__(self, pars=None, **kwargs):
+        super().__init__(name='msm')
+        self.define_pars(
+            msm_share=ss.bernoulli(p=0.015),
+        )
+        self.update_pars(pars=pars, **kwargs)
+
+        return
+
+    def set_network_states(self, upper_age=None):
+        self.set_msm(upper_age=upper_age)
+        return
+
+    def set_msm(self, upper_age=None):
+        _, m_uids = self._get_uids(upper_age=upper_age)
+        self.participant[m_uids] = self.pars.msm_share.rvs(m_uids)
+        return
+
+    def match_pairs(self):
+        """ Match males by age using sorting """
+        ppl = self.sim.people
+
+        # Find people eligible for a relationship
+        active = self.over_debut
+        underpartnered = self.partners < self.concurrency
+        m_eligible = active & ppl.male & underpartnered
+        m_looking = self.pars.p_pair_form.filter(m_eligible.uids)
+
+        if len(m_looking) == 0:
+            raise NoPartnersFound()
+
+        # Match mairs by sorting the men looking for partners by age, then matching pairs by taking
+        # 2 people at a time from the sorted list
+        m_ages = ppl.age[m_looking]
+        ind_m = np.argsort(m_ages)
+        p1 = m_looking[ind_m][::2]
+        p2 = m_looking[ind_m][1::2]
+        maxlen = min(len(p1), len(p2))
+        p1 = p1[:maxlen]
+        p2 = p2[:maxlen]
+
+        # Make sure everyone only appears once (?)
+        if len(np.intersect1d(p1, p2)):
+            errormsg = 'Some people appear in both p1 and p2'
+            raise ValueError(errormsg)
+
+        return p1, p2
+
+
+class AgeApproxMSM(MFNetwork):
+    """Men-who-have-sex-with-men network using approximate age-preference matching.
+
+    Extends :class:`StructuredSexual` for MSM partnerships. Unlike
+    :class:`AgeMatchedMSM`, this variant splits eligible males into two
+    arbitrary groups and matches them using the standard age-difference
+    preference distributions rather than exact age sorting.
+
+    Args:
+        **kwargs: Parameter overrides forwarded to :class:`StructuredSexual`.
+    """
+
+    def __init__(self, **kwargs):
+        super().__init__(name='msm', **kwargs)
+
+    def match_pairs(self, ppl):
+        """ Match pairs using age preferences """
+
+        # Find people eligible for a relationship
+        active = self.over_debut()
+        underpartnered = self.partners < self.concurrency
+        m_eligible = active & ppl.male & underpartnered
+        m_looking = self.pars.p_pair_form.filter(m_eligible.uids)
+
+        # Split the total number of males looking for partners into 2 groups
+        # The first group will be matched with the second group
+        group1 = m_looking[::2]
+        group2 = m_looking[1::2]
+        loc, scale = self.get_age_risk_pars(group1, self.pars.age_diff_pars)
+        self.pars.age_diffs.set(loc=loc, scale=scale)
+        age_gaps = self.pars.age_diffs.rvs(group1)
+        desired_ages = ppl.age[group1] + age_gaps
+        g2_ages = ppl.age[group2]
+        ind_p1 = np.argsort(g2_ages)
+        ind_p2 = np.argsort(desired_ages)
+        p1 = m_eligible.uids[ind_p1]
+        p2 = group2[ind_p2]
+        maxlen = min(len(p1), len(p2))
+        p1 = p1[:maxlen]
+        p2 = p2[:maxlen]
+
+        return p1, p2
